First version of application management commands - `init`, `refresh` and `destroy` are ready.
