+++ clustered/config/Environment_Config.json
@@ -1,7 +1,7 @@
 {
     "DB_CONFIG": {
         "DB_ENGINE": "sqlite",
-        "DB_FILE": "/Volumes/WorkSpace/POC/SparkClusterEC2/sparkclusteronaws/clustered/database/clustered.db"
+        "DB_FILE": "/Volumes/WorkSpace/POC/Mjolnir/clustered/db/clustered.db"
     },
     "USER_CONFIG": [
         {
