--- clustered/database/__init__.py
@@ -1,3 +1,0 @@
-from . import database_engine
-
-db_obj = database_engine.DatabaseEngine()
